Reorganise code to make it work
